requirements.txt
flask==3.1.0
flasgger==0.9.7.1
requests==2.32.3
psycopg2-binary==2.9.10
python-dotenv==1.0.1